Make Dispatcher::backendFallbackKernels_ an array
We already made OperatorEntry::dispatchTable_ an array to be able to avoid the concurrency primitives there,
but Dispatcher::backendFallbackKernels_ has the same issue. Let's make it a table too.

Since there is some code duplication here, we also factor out the concept of a KernelFunctionTable to be used in both places.

Differential Revision: [D18663426](https://our.internmc.facebook.com/intern/diff/D18663426/)

ghstack-source-id: 94449179
Pull Request resolved: https://github.com/pytorch/pytorch/pull/30340